requirements.txt
# Core Dependencies
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
redis==5.0.2
python-dotenv==1.0.1

# Data Processing & Analysis
pandas==2.2.1
numpy==1.26.4
scipy==1.12.0
statsmodels==0.14.1
scikit-learn==1.4.1

# Visualization
plotly==5.19.0
matplotlib==3.8.3
seaborn==0.13.2

# LLM Integration
openai==1.12.0
anthropic==0.18.1
tiktoken==0.6.0
tenacity==8.2.3

# Testing
pytest==8.0.2
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.5

# Code Quality
black==24.2.0
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# Development Tools
python-json-logger==2.0.7
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
aiohttp==3.9.3
uvicorn==0.27.1
gunicorn==21.2.0

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
myst-parser==2.0.0 